backend/requirements.txt
fastapi
uvicorn[standard]
google-adk
requests
python-dotenv
pydantic